+++ modules/channel-vonage/src/backend/client.ts
@@ -19,7 +19,9 @@ import {
   VonageAudioRenderer,
   VonageMediaTemplateRenderer,
   VonageLocationRenderer,
-  VonageTemplateRenderer
+  VonageTemplateRenderer,
+  VonageDropdownRenderer,
+  VonageFileRenderer
 } from '../renderers'
 import { VonageCommonSender, VonageTypingSender } from '../senders'
 
@@ -105,6 +107,7 @@ export class VonageClient {
 
     this.renderers = [
       new VonageCardRenderer(),
+      new VonageDropdownRenderer(),
       new VonageTextRenderer(),
       new VonageImageRenderer(),
       new VonageLocationRenderer(),
@@ -113,6 +116,7 @@ export class VonageClient {
       new VonageVideoRenderer(),
       new VonageTemplateRenderer(),
       new VonageMediaTemplateRenderer(),
+      new VonageFileRenderer(),
       new VonageChoicesRenderer()
     ]