+++ src/bp/core/config/botpress.config.ts
@@ -217,7 +217,7 @@ export interface BotpressConfig {
         title?: string
         /**
          * Replace the default favicon
-         * @example "assets/ui-studio/public/img/favicon.png"
+         * @example "assets/studio/ui/public/img/favicon.png"
          */
         favicon?: string
         /**
@@ -281,7 +281,7 @@ export interface BotpressConfig {
     maxFileSize: string
     /**
      * The list of allowed extensions for media file uploads
-     * @default ["image/jpeg","image/png","image/gif","audio/mpeg","video/mp4"]
+     * @default ["image/jpeg","image/png","image/gif","audio/mpeg","video/mp4","application/pdf"]
      */
     allowedMimeTypes: string[]
   }