+++ src/bp/core/modules/config-reader.ts
@@ -1,9 +1,11 @@
 import { Logger, ModuleEntryPoint } from 'botpress/sdk'
+import { ObjectCache } from 'common/object-cache'
 import { GhostService } from 'core/bpfs'
 import fs from 'fs'
 import defaultJsonBuilder from 'json-schema-defaults'
 import _ from 'lodash'
 import { Memoize } from 'lodash-decorators'
+import LRU from 'lru-cache'
 import path from 'path'
 import { VError } from 'verror'
 import yn from 'yn'
@@ -23,14 +25,22 @@ export type ModuleConfig = Dic<any>
  */
 export class ConfigReader {
   private modules = new Map<string, ModuleEntryPoint>()
+  private moduleConfigCache: LRU<string, any>
 
-  constructor(private logger: Logger, modules: ModuleEntryPoint[], private ghost: GhostService) {
+  constructor(
+    private logger: Logger,
+    modules: ModuleEntryPoint[],
+    private ghost: GhostService,
+    private cache: ObjectCache
+  ) {
     for (const module of modules) {
       const name = _.get(module, 'definition.name', '').toLowerCase()
       if (name.length) {
         this.modules.set(name, module)
       }
     }
+    this.moduleConfigCache = new LRU()
+    this._listenForModuleConfigCacheInvalidation()
   }
 
   public async initialize() {
@@ -40,19 +50,25 @@ export class ConfigReader {
   @Memoize()
   private async getModuleConfigSchema(moduleId: string): Promise<any> {
     const modulePath = process.LOADED_MODULES[moduleId]
-    const configSchema = path.resolve(modulePath, 'assets', 'config.schema.json')
+    if (!modulePath) {
+      this.logger.warn(`Cannot load the config schema for module "${moduleId}". Module is disabled.`)
+      return {}
+    }
 
     try {
+      const configSchema = path.resolve(modulePath, 'assets', 'config.schema.json')
+
       if (fs.existsSync(configSchema)) {
         return JSON.parse(fs.readFileSync(configSchema, 'utf-8'))
       }
     } catch (err) {
       this.logger.attachError(err).error(`Error while loading the config schema for module "${moduleId}"`)
     }
+
     return {}
   }
 
-  public async loadFromDefaultValues(moduleId) {
+  public async loadFromDefaultValues(moduleId: string) {
     return defaultJsonBuilder(await this.getModuleConfigSchema(moduleId))
   }
 
@@ -83,7 +99,7 @@ export class ConfigReader {
 
     /* START DEPRECATED */
     // TODO: Remove support for those old env variables in BP 12 (we need to add those to 11 -> 12 migration guide)
-    for (const option of Object.keys(schema.properties || {})) {
+    for (const option of Object.keys(schema?.properties || {})) {
       const keyOld = `BP_${moduleId}_${option}`.toUpperCase()
       if (keyOld in process.env) {
         const keyNew = `BP_MODULE_${moduleId}_${option}`.toUpperCase()
@@ -111,7 +127,7 @@ export class ConfigReader {
   }
 
   @Memoize()
-  private async getModuleDefaultConfigFile(moduleId): Promise<any | undefined> {
+  private async getModuleDefaultConfigFile(moduleId: string): Promise<any | undefined> {
     try {
       const defaultConfig = {
         $schema: `../../assets/modules/${moduleId}/config.schema.json`,
@@ -172,21 +188,41 @@ export class ConfigReader {
     return config
   }
 
-  @Memoize()
-  public async getGlobal(moduleId: string): Promise<ModuleConfig> {
-    return this.getMerged(moduleId)
+  private async getCachedOrFresh(key: string) {
+    if (this.moduleConfigCache.has(key)) {
+      return this.moduleConfigCache.get(key) || {}
+    }
+
+    const [moduleId, botId, ignoreGlobal] = key.split('//')
+    const config = await this.getMerged(moduleId, botId, yn(ignoreGlobal))
+    this.moduleConfigCache.set(key, config)
+
+    return config
   }
 
-  // Don't @Memoize() this fn. It only memoizes on the first argument
-  // https://github.com/steelsojka/lodash-decorators/blob/master/src/memoize.ts#L15
+  public async getGlobal(moduleId: string): Promise<ModuleConfig> {
+    return this.getCachedOrFresh(moduleId)
+  }
+
   public getForBot(moduleId: string, botId: string, ignoreGlobal?: boolean): Promise<ModuleConfig> {
     const cacheKey = `${moduleId}//${botId}//${!!ignoreGlobal}`
-    return this.getForBotMemoized(cacheKey)
+    return this.getCachedOrFresh(cacheKey)
   }
 
-  @Memoize()
-  public getForBotMemoized(cacheKey: string): Promise<ModuleConfig> {
-    const [moduleId, botId, ignoreGlobal] = cacheKey.split('//')
-    return this.getMerged(moduleId, botId, yn(ignoreGlobal))
+  private _listenForModuleConfigCacheInvalidation() {
+    this.cache.events.on('invalidation', key => {
+      try {
+        const moduleId = key.match(/^.*::data\/.*\/config\/([\s\S]+([a-zA-Z0-9-_])+\.json)/i)?.[1]?.replace('.json', '')
+
+        if (moduleId) {
+          this.moduleConfigCache
+            .keys()
+            .filter(x => x.startsWith(moduleId))
+            .forEach(x => this.moduleConfigCache.del(x))
+        }
+      } catch (err) {
+        this.logger.error('Error invalidating module config cache: ' + err.message)
+      }
+    })
   }
 }