+++ src/bp/migrations/v12_21_1-1618260400-bot_content_type_update.ts
@@ -5,15 +5,14 @@ import _ from 'lodash'
 const migration: Migration = {
   info: {
     description: 'Adds video and audio content types to bot configs',
-    target: 'core',
+    target: 'bot',
     type: 'config'
   },
-  up: async ({ bp, configProvider }: MigrationOpts): Promise<sdk.MigrationResult> => {
-    const bots = await bp.bots.getAllBots()
-    const newTypes = ['builtin_video', 'builtin_audio']
+  up: async ({ bp, configProvider, metadata }: MigrationOpts): Promise<sdk.MigrationResult> => {
     let hasChanges = false
 
-    for (const [botId, botConfig] of bots) {
+    const updateBotContentTypes = async (botId: string, botConfig: sdk.BotConfig) => {
+      const newTypes = ['builtin_video', 'builtin_audio']
       const { contentTypes } = botConfig.imports
 
       // Fix for the previous migration which was removed
@@ -28,7 +27,21 @@ const migration: Migration = {
       }
     }
 
+    if (metadata.botId) {
+      const botConfig = await bp.bots.getBotById(metadata.botId)
+      await updateBotContentTypes(metadata.botId, botConfig!)
+    } else {
+      const bots = await bp.bots.getAllBots()
+      for (const [botId, botConfig] of bots) {
+        await updateBotContentTypes(botId, botConfig)
+      }
+    }
+
     return { success: true, message: hasChanges ? 'New types added successfully' : 'Nothing to change' }
+  },
+  down: async () => {
+    // Down migrations not necessary for content types, no impact
+    return { success: true, message: 'Nothing to change' }
   }
 }